# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Community Services Marketplace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: meta
  - name: providers
  - name: jobs
  - name: reviews
  - name: ads
  - name: favorites
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /meta:
    get:
      operationId: getMeta
      tags: [meta]
      summary: Countries, locations and service categories
      responses:
        "200":
          description: Meta info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
  /home:
    get:
      operationId: getHomeSummary
      tags: [meta]
      summary: Homepage summary content
      responses:
        "200":
          description: Home summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeSummary"
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: Search providers
      parameters:
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: countryId
          in: query
          schema: { type: integer }
        - name: cityId
          in: query
          schema: { type: integer }
        - name: communityId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: verifiedOnly
          in: query
          schema: { type: boolean }
        - name: sort
          in: query
          schema:
            type: string
            enum: [value, rating, reviews, newest]
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSummary"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Provider profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Provider detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /providers/{id}/reviews:
    get:
      operationId: listProviderReviews
      tags: [reviews]
      summary: Reviews for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createProviderReview
      tags: [reviews]
      summary: Submit a review for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Browse community jobs
      parameters:
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: countryId
          in: query
          schema: { type: integer }
        - name: cityId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: mine
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobSummary"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Job detail with expressions of interest
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /jobs/{id}/interests:
    post:
      operationId: createJobInterest
      tags: [jobs]
      summary: Express interest in a job (provider)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInterestInput"
      responses:
        "201":
          description: Created interest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobInterest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /jobs/{id}/select:
    post:
      operationId: selectJobProvider
      tags: [jobs]
      summary: Customer selects the winning provider
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderSelection"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /jobs/{id}/status:
    patch:
      operationId: updateJobStatus
      tags: [jobs]
      summary: Update job status (e.g. mark completed or cancelled)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobStatusUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: Community advertisements
      parameters:
        - name: countryId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Advertisement"
  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: Saved providers
      responses:
        "200":
          description: Saved providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSummary"
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Save a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /favorites/{providerId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a saved provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]
    ApiMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
    Country:
      type: object
      required: [id, name, code, currencyCode, currencySymbol, flagEmoji, cities]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        currencyCode: { type: string }
        currencySymbol: { type: string }
        flagEmoji: { type: string }
        cities:
          type: array
          items:
            $ref: "#/components/schemas/City"
    City:
      type: object
      required: [id, name, region, communities]
      properties:
        id: { type: integer }
        name: { type: string }
        region: { type: string }
        communities:
          type: array
          items:
            $ref: "#/components/schemas/Community"
    Community:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    ServiceCategory:
      type: object
      required: [id, name, groupName, icon]
      properties:
        id: { type: integer }
        name: { type: string }
        groupName: { type: string }
        icon: { type: string }
        providerCount: { type: integer }
    Meta:
      type: object
      required: [countries, categories]
      properties:
        countries:
          type: array
          items:
            $ref: "#/components/schemas/Country"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ServiceCategory"
    HomeSummary:
      type: object
      required: [popularCategories, topProviders, bestValueProviders, recentJobs, ads, stats]
      properties:
        popularCategories:
          type: array
          items:
            $ref: "#/components/schemas/ServiceCategory"
        topProviders:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
        bestValueProviders:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
        recentJobs:
          type: array
          items:
            $ref: "#/components/schemas/JobSummary"
        ads:
          type: array
          items:
            $ref: "#/components/schemas/Advertisement"
        stats:
          $ref: "#/components/schemas/MarketplaceStats"
    MarketplaceStats:
      type: object
      required: [providerCount, verifiedProviderCount, openJobCount, reviewCount, communityCount]
      properties:
        providerCount: { type: integer }
        verifiedProviderCount: { type: integer }
        openJobCount: { type: integer }
        reviewCount: { type: integer }
        communityCount: { type: integer }
    ProviderSummary:
      type: object
      required:
        [id, businessName, providerType, tagline, categories, communityName,
         cityName, countryId, currencySymbol, verificationStatus, avgRating,
         reviewCount, valueScore, isFavorite]
      properties:
        id: { type: integer }
        businessName: { type: string }
        providerType:
          type: string
          enum: [individual, sole_trader, small_business, company]
        tagline: { type: string }
        categories:
          type: array
          items: { type: string }
        communityName: { type: string }
        cityName: { type: string }
        countryId: { type: integer }
        currencySymbol: { type: string }
        verificationStatus:
          type: string
          enum: [not_verified, pending, verified, rejected, suspended]
        avgRating: { type: number }
        reviewCount: { type: integer }
        valueScore: { type: integer }
        valueScoreExplanation: { type: ["string", "null"] }
        priceIndicator: { type: ["string", "null"] }
        yearsActive: { type: ["integer", "null"] }
        isFavorite: { type: boolean }
    ProviderDetail:
      allOf:
        - $ref: "#/components/schemas/ProviderSummary"
        - type: object
          required: [description, serviceAreas, ratingBreakdown, reviews]
          properties:
            description: { type: string }
            serviceAreas:
              type: array
              items: { type: string }
            ratingBreakdown:
              $ref: "#/components/schemas/RatingBreakdown"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            completedJobCount: { type: integer }
    RatingBreakdown:
      type: object
      required: [quality, price, reliability, professionalism, overall]
      properties:
        quality: { type: number }
        price: { type: number }
        reliability: { type: number }
        professionalism: { type: number }
        overall: { type: number }
    Review:
      type: object
      required:
        [id, providerId, reviewerName, comment, verifiedJob, createdAt,
         quality, price, reliability, professionalism, overall]
      properties:
        id: { type: integer }
        providerId: { type: integer }
        reviewerName: { type: string }
        comment: { type: string }
        verifiedJob: { type: boolean }
        createdAt: { type: string }
        quality: { type: integer }
        price: { type: integer }
        reliability: { type: integer }
        professionalism: { type: integer }
        overall: { type: integer }
        providerResponse: { type: ["string", "null"] }
    ReviewInput:
      type: object
      required: [reviewerName, comment, quality, price, reliability, professionalism, overall]
      properties:
        reviewerName: { type: string, minLength: 1 }
        comment: { type: string, minLength: 1 }
        quality: { type: integer, minimum: 1, maximum: 5 }
        price: { type: integer, minimum: 1, maximum: 5 }
        reliability: { type: integer, minimum: 1, maximum: 5 }
        professionalism: { type: integer, minimum: 1, maximum: 5 }
        overall: { type: integer, minimum: 1, maximum: 5 }
    JobSummary:
      type: object
      required:
        [id, title, categoryName, categoryIcon, communityName, cityName,
         countryId, currencySymbol, budgetType, status, urgency, postedAt,
         interestCount, isMine]
      properties:
        id: { type: integer }
        title: { type: string }
        categoryName: { type: string }
        categoryIcon: { type: string }
        communityName: { type: string }
        cityName: { type: string }
        countryId: { type: integer }
        currencySymbol: { type: string }
        budgetType:
          type: string
          enum: [fixed, range, open]
        budgetMin: { type: ["number", "null"] }
        budgetMax: { type: ["number", "null"] }
        status:
          type: string
          enum: [draft, open, providers_interested, provider_selected, in_progress, completed, reviewed, cancelled, expired, disputed]
        urgency:
          type: string
          enum: [flexible, soon, urgent, emergency]
        postedAt: { type: string }
        preferredDate: { type: ["string", "null"] }
        interestCount: { type: integer }
        isMine: { type: boolean }
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/JobSummary"
        - type: object
          required: [description, interests]
          properties:
            description: { type: string }
            selectedInterestId: { type: ["integer", "null"] }
            interests:
              type: array
              items:
                $ref: "#/components/schemas/JobInterest"
    JobInput:
      type: object
      required: [title, description, categoryId, communityId, budgetType, urgency]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        categoryId: { type: integer }
        communityId: { type: integer }
        budgetType:
          type: string
          enum: [fixed, range, open]
        budgetMin: { type: ["number", "null"] }
        budgetMax: { type: ["number", "null"] }
        urgency:
          type: string
          enum: [flexible, soon, urgent, emergency]
        preferredDate: { type: ["string", "null"] }
    JobInterest:
      type: object
      required:
        [id, jobId, providerId, providerName, verificationStatus, avgRating,
         reviewCount, valueScore, message, canMeetBudget, createdAt, currencySymbol]
      properties:
        id: { type: integer }
        jobId: { type: integer }
        providerId: { type: integer }
        providerName: { type: string }
        verificationStatus:
          type: string
          enum: [not_verified, pending, verified, rejected, suspended]
        avgRating: { type: number }
        reviewCount: { type: integer }
        valueScore: { type: integer }
        message: { type: string }
        canMeetBudget: { type: boolean }
        estimateMin: { type: ["number", "null"] }
        estimateMax: { type: ["number", "null"] }
        availability: { type: ["string", "null"] }
        createdAt: { type: string }
        currencySymbol: { type: string }
    JobInterestInput:
      type: object
      required: [providerId, message, canMeetBudget]
      properties:
        providerId: { type: integer }
        message: { type: string, minLength: 1 }
        canMeetBudget: { type: boolean }
        estimateMin: { type: ["number", "null"] }
        estimateMax: { type: ["number", "null"] }
        availability: { type: ["string", "null"] }
    ProviderSelection:
      type: object
      required: [interestId]
      properties:
        interestId: { type: integer }
    JobStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, in_progress, completed, reviewed, cancelled, disputed]
    Advertisement:
      type: object
      required: [id, businessName, title, description, countryId, targetArea, imageUrl]
      properties:
        id: { type: integer }
        businessName: { type: string }
        title: { type: string }
        description: { type: string }
        countryId: { type: integer }
        targetArea: { type: string }
        imageUrl: { type: string }
        linkUrl: { type: ["string", "null"] }
    FavoriteInput:
      type: object
      required: [providerId]
      properties:
        providerId: { type: integer }